+++ pleco/src/bots/minimax.rs
@@ -21,7 +21,7 @@ pub fn minimax(board: &mut Board, max_depth: u16) -> BestMove {
     let moves = board.generate_moves();
     if moves.is_empty() {
         if board.in_check() {
-            return BestMove::new_none(MATE + (board.depth() as i16));
+            return BestMove::new_none(MATE + (board.depth() as i32));
         } else {
             return BestMove::new_none(DRAW);
         }
